LoginPage V1.1
Used non-blurry image for background, but easily changed if desired.  "Go Forth" marquee a little too big, so we should try to figure out how to resize it to a more appropriate proportion.  Applied a few different colors underneath the input forms, we need to settle on which colors we want to use for those.  The inputs themselves could probably be a different color as well.